requirements.txt
ansible
pybatfish
pytest
PyYAML